Fix algorithm in case of packet loss in the last window.
If the sender sends the last block, last_block is set from -1 to the
block number of the last block.  However, if a datagram gets lost in
that last window, this is not correct as at least part of the window
has to be sent again.  The receiver complains with an ACK of the last
block received in sequence instead of the ACK of the last block.
Reset last_block to -1 in that case to properly repeat sending the
remaining blocks of the last window.

Signed-off-by: Martin Dummer <md11@users.sourceforge.net>